Add Image model for storing uploaded images
This commit adds a new Django model called Image, which is used to store images uploaded by users. The model has a single field named `image`, which is an ImageField that stores the uploaded image files in the 'images/' directory.

This model will be used in conjunction with other features that require uploading and storing images, such as user profile pictures or image predictions.